[MA-26] feat : 이용내역 조회 -> 상세 조회 nav 연결

+++ domain/src/main/java/com/motax/modutaxi/domain/model/MonthlyUsageHistoryData.kt
@@ -9,7 +9,7 @@ data class MonthlyUsageHistoryData(
 )
 
 data class UsageHistoryDataItem(
-    val historyId: Int,
+    val historyId: Long,
     val departureTime: String,
     val departureName: String,
     val arrivalName: String,
